Handle SMS and eMail notification sub-actuators
Existing logic failed to recognize SMS and eMail notifications as sub-actuators of the generic actuator. The result was an exception being thrown.
I patched the logic so that the state is set to on if there are notifications remaining and to off, if there are no notifications remaining. This should match how it is displayed in the GUI.